+++ packages/auth/graphql/resolvers/_queries/user.js
@@ -1,17 +1,13 @@
 const transformUser = require('../../../lib/transformUser')
+const { resolveUser } = require('../../../lib/Users')
 const Roles = require('../../../lib/Roles')
-const isUUID = require('is-uuid')
 
 module.exports = async (_, args, { pgdb, user: me }) => {
   const { slug } = args
   if (!slug) {
     return null
   }
-  const user = await pgdb.public.users.findOne(
-    isUUID.v4(slug)
-      ? {id: slug}
-      : {username: slug}
-  )
+  const user = await resolveUser({ slug, pgdb, fallback: me })
   if (
     user &&
     (user.hasPublicProfile || Roles.userIsMeOrInRoles(user, me, ['admin', 'supporter']))
